refactor(github-issues-runtime): split runtime helper modules (#1256)
